requirements.txt
scrapy
python-dotenv
environ